[FrameworkBundle][Routing] Deprecate annotations
