+++ src/Symfony/Component/Routing/CHANGELOG.md
@@ -8,6 +8,9 @@ CHANGELOG
  * Add native return type to `AnnotationClassLoader::setResolver()`
  * Deprecate Doctrine annotations support in favor of native attributes
  * Change the constructor signature of `AnnotationClassLoader` to `__construct(?string $env = null)`, passing an annotation reader as first argument is deprecated
+ * Deprecate `AnnotationClassLoader`, use `AttributeClassLoader` instead
+ * Deprecate `AnnotationDirectoryLoader`, use `AttributeDirectoryLoader` instead
+ * Deprecate `AnnotationFileLoader`, use `AttributeFileLoader` instead
 
 6.2
 ---
