+++ src/Symfony/Component/Routing/Loader/AnnotationDirectoryLoader.php
@@ -11,73 +11,15 @@
 
 namespace Symfony\Component\Routing\Loader;
 
-use Symfony\Component\Config\Resource\DirectoryResource;
-use Symfony\Component\Routing\RouteCollection;
+trigger_deprecation('symfony/routing', '6.4', 'The "%s" class is deprecated, use "%s" instead.', AnnotationDirectoryLoader::class, AttributeDirectoryLoader::class);
 
-/**
- * AnnotationDirectoryLoader loads routing information from annotations set
- * on PHP classes and methods.
- *
- * @author Fabien Potencier <fabien@symfony.com>
- */
-class AnnotationDirectoryLoader extends AnnotationFileLoader
-{
+class_exists(AttributeDirectoryLoader::class);
+
+if (false) {
     /**
-     * @throws \InvalidArgumentException When the directory does not exist or its routes cannot be parsed
+     * @deprecated since Symfony 6.4, to be removed in 7.0, use {@link AttributeDirectoryLoader} instead
      */
-    public function load(mixed $path, string $type = null): ?RouteCollection
+    class AnnotationDirectoryLoader
     {
-        if (!is_dir($dir = $this->locator->locate($path))) {
-            return parent::supports($path, $type) ? parent::load($path, $type) : new RouteCollection();
-        }
-
-        $collection = new RouteCollection();
-        $collection->addResource(new DirectoryResource($dir, '/\.php$/'));
-        $files = iterator_to_array(new \RecursiveIteratorIterator(
-            new \RecursiveCallbackFilterIterator(
-                new \RecursiveDirectoryIterator($dir, \FilesystemIterator::SKIP_DOTS | \FilesystemIterator::FOLLOW_SYMLINKS),
-                fn (\SplFileInfo $current) => !str_starts_with($current->getBasename(), '.')
-            ),
-            \RecursiveIteratorIterator::LEAVES_ONLY
-        ));
-        usort($files, fn (\SplFileInfo $a, \SplFileInfo $b) => (string) $a > (string) $b ? 1 : -1);
-
-        foreach ($files as $file) {
-            if (!$file->isFile() || !str_ends_with($file->getFilename(), '.php')) {
-                continue;
-            }
-
-            if ($class = $this->findClass($file)) {
-                $refl = new \ReflectionClass($class);
-                if ($refl->isAbstract()) {
-                    continue;
-                }
-
-                $collection->addCollection($this->loader->load($class, $type));
-            }
-        }
-
-        return $collection;
-    }
-
-    public function supports(mixed $resource, string $type = null): bool
-    {
-        if (!\is_string($resource)) {
-            return false;
-        }
-
-        if (\in_array($type, ['annotation', 'attribute'], true)) {
-            return true;
-        }
-
-        if ($type) {
-            return false;
-        }
-
-        try {
-            return is_dir($this->locator->locate($resource));
-        } catch (\Exception) {
-            return false;
-        }
     }
 }
