+++ src/Symfony/Component/Routing/Tests/Fixtures/AttributedClasses/BazClass.php
@@ -9,8 +9,11 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Component\Routing\Tests\Fixtures\AnnotatedClasses;
+namespace Symfony\Component\Routing\Tests\Fixtures\AttributedClasses;
 
-class FooClass
+class BazClass
 {
+    public function __invoke()
+    {
+    }
 }
