+++ src/Symfony/Component/Routing/Tests/Fixtures/AttributedClasses/FooTrait.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Symfony\Component\Routing\Tests\Fixtures\AnnotatedClasses;
+namespace Symfony\Component\Routing\Tests\Fixtures\AttributedClasses;
 
 trait FooTrait
 {
